requirements.txt
streamlit
langchain
langchain-community
langchain-chroma
langchain-groq
langchain-huggingface
python-dotenv
pypdf
pandas
huggingface-hub
sentence-transformers
torch
chromadb
gspread
google-auth
beautifulsoup4
requests

